functions/generate-combined-newsletter/requirements.txt
functions-framework==3.*
google-cloud-bigquery>=3.10.0
google-cloud-aiplatform==1.71.1